Summary: add history for tracking of changes
Details:
fixes https://github.com/redhat-performance/ops-tools/issues/15

Specifically this adds a "history" section to the yaml data file
that tracks each host.  The format of the the history section is:

history:
  hostname1.example.com:
    0: cloudname1
    1476203752: cloudname2

and so forth.  The initial definition of a host will create the
history with the start time of 0 (epoch) indicating which cloud env
the host belongs to.  Any scheduled overrides, or even move of the
default location of host will add a new history entry with unix
time (seconds since epoch) indicating the host is moved.

Deleting the host will not delete the history.  A host may be deleted
and added again at a later time, and the history will still reflect
environment membership.

This allows for calendar rendering that reflects the actual usage
for all host resources.