WIP: add config options
Details:
add a yaml file with config options.  This is a
first pass to make the entire quads package portable
and usable outside of our own environment.  The various
scripts will need to be updated to read in the yaml from
the config source (for both shell and python elements)

Also adding the csv-to-instack.py used to generate instackenv.json

This is related to: https://github.com/redhat-performance/quads/issues/1

Change-Id: I6c052fe46617f3fbbdae8c9f0b1480d477fc1c87