Summary: massive updates to make portable
Details:
use the settings in the conf/quads.yml file.
There is now a mechanism for having either python
or bash scripts read the yaml config file. This is
done via the yay.sh.  We wrap this up using
bin/load-config.sh

quads should be much more portable now and the only
changes needed are in the conf/quads.yml file.  This
of course will need to be tested and validated.

Change-Id: I5aafcc43b42798a1ff60af59e547fc5d537b7912