Summary: simple html generator
Details:
Run ./simple-table-generator.sh 2016-11 30
This script in turn also calls ./simple-table-generator.py which
spits out a simple HTML file that can show a colorized lab usage
for the month passed as an argument.  Currently a little slow
due to the large number of hosts and that each day is viewed for
the full month (arg2) per each of the hosts.  A possible speed up
will be to move much of quads functionality into libraries that can
then be included (as opposed to an external shell call).

Change-Id: I30690105a22390fea3a85233c074fbfad561f75f