[WIP] notifications for cloud schedule starting
Taking a new approach to notifications for when an
assignment starts and we want to optionally notify the admins.

This will use the UNIX 'at' command to schedule a one-time job
to send an email at the time matching --schedule-start via quads.

We want this optional because we don't always need to know.

e.g. bin/schedule-at-notify-cloud-start.py -d "2016-11-22" -c cloud09

Issues:

* 'at' doesn't seem to take STDIN variables like $1 and $2 when
  you pass it a shell script.

Possible workarounds:

* We could have Python generate a temporary template containing
  the -c (cloud) variable and then send the message.

Improvements:

* Add the necessary netcat commands to communicate with IRC bot
  to notify IRC channel when a schedule starts.

related to: https://github.com/redhat-performance/quads/issues/2

Change-Id: I2f15effb48e90b884dd9ebe0111426c8ef00c98a