Summary: simplify juniper switch commands
Details:
Juniper suggested that the double commit might
be an issue.  This change fixes the double commit
ensuring a consistent configuration when the changes
are activated rather than changing the interface
settings and then doing a second commit to change the vlan
interface changes.

It also simplifies the command set (as it's not necessary
to drill down into the vlans section before changing it).
The vlans section of the config can be changed similarly
to the interfaces section (from the top level)

Change-Id: I308feb12d216883ab62a669fca7bb0607f2c3be8