dont run the wiki regen from move command
The wiki regeneration is currently too costly to run
via the move command.  Having that call there slows down
the process of moving a large number of hosts.  Just
run the wiki regen more frequenty via cron.  Since it
checks a lock file for another process already running
it is OK to run it more frequently.

Also reduce the sleep for the ipmi poweroff/poweron.

Change-Id: I7bef2a7f938b7a16ff0dfc11c0026be961c1080d