critical fix.
the boot order check was being terminated prematurely if
any one of the hosts was still running a playbook.  Also
there appears to be a setting in the director deployed hosts
that prevents the playbook from "resetting" the host, so
a powercycle is added to ensure components of dell systems
can be reconfigured.

Change-Id: I4578f3194f4db8a46d8f9de15fc9dbf2562ebc01