Critical fix.
the boot order check was being terminated prematurely if
any one of the hosts was still running a playbook.  Also
there appears to be a setting in the director deployed hosts
that prevents the playbook from "resetting" the host, so
a powercycle is added to ensure components of dell systems
can be reconfigured.

fixes https://github.com/redhat-performance/quads/issues/30

Change-Id: I25c16afe4dccc6647061f513950cfc51fb6be007