new qinq modes for environments
The interfaces wired to each host are defined in the
$data_dir/ports/<fqdn>.  The previous mode was to have
vlan separation across all the interfaces.  e.g. if you
have 4 interfaces on a host, you would have 4 separate VLANs
and each interface would be a member of a QinQ enabled VLAN
on the connected switch.

A new yaml key value pair called "qinq" is defined under the
cloud definition in order to set the mode in which the host
interfaces are configured.  The existing mode is:

   qinq: 0   (default / current behavior)

if this key is not set as part of the cloud definition, then
it will default to 0, which should cause VLAN separation across
all the interfaces.

   qinq: 1  (combine all interfaces into the same VLAN)

if set to 1, then all the interfaces will use the same VLAN
as the VLAN id allocated for em1 (the first interface) of the
environment.

Two new options have been added to read and set the flag:

   quads --ls-qinq [ ... ]  (queries the current value if set)
   quads --define-cloud [...] --qinq <value>

Note, that if you set the qinq value for an cloud environment,
and then redefine the env (e.g. when re-allocating it to a new
tenant), you will either need to set this flag, or if not explicitly
set will default back to 0 on the re-assignment.

Fixes: https://github.com/redhat-performance/quads/issues/32

Change-Id: I592757fb1b954d18d110d3d06fc2e614a1d8fefe