Use dedicated IPMI acct, reset on new environment
This switches a few things over to using a dedicated
IPMI account for users and their environments, it also
changes the IPMI account password so that they aren't
the same (defaults to ticket number).

This will also be reflected in the generated instackenv.json
so that within each cloud/environment the IPMI password
is different.

fixes https://github.com/redhat-performance/quads/issues/33

Change-Id: Ieb86ec289faac3b0be249881319e03dce3fd7024