allow for designation of the undercloud
This commit is intended to give the users freedom to set
the hosts that are excluded from the instackenv.json file.
We leverage the foreman_director_parameter option that can
be set in foreman.  The behavior is by default (during initial
provisioning) to have the first host in the environment have

$foreman_director_parameter == false

and all other hosts to have the parameter set to true.  e.g.
if you define: $foreman_director_parameter = "undercloud",
then you can go into foreman and change the host(s) that
have undercloud=false.  Any host(s) that have undercloud set
to true will then not be included in the instackenv.json.

This is useful if you want to pick your own undercloud or
otherwise exclude a host or a set of hosts from your openstack
deployment to be used for other purposes.

Change-Id: I13aa5358cbabc99082ef7a4708769932d1149600
Fixes: https://github.com/redhat-performance/quads/issues/31