Logic error in computing schedule index
There is currently a logic bug in lib/quads.py that makes some
assumptions that the next available assignment id will always
be iterative. This can be triggered when a schedule has been
removed and another net-new schedule is added. This is a serious
issue that can cause disruption/reprovisioning of an existing
environment.

The length of the array of keys in the schedule dictionary was
assumed to be the correct next index.  This assumption breaks
when a host has schedules that were removed.

This hot-fix corrects this problem.

Change-Id: I7bc87cd4b42c222954215c334e4de4e97fca0e93
Fixes: https://github.com/redhat-performance/quads/issues/35