Defined cloud range loaded via libquads
Currently simple-table-generator.py uses a range within
the tool to define the in-use start/end clouds in order to
generate heat map visualizations.  When you exceed this number
of defined clouds it breaks the generation.

This will load and query the QUADS object instead and obtain
the range of defined clouds so we know what to generate the
visualizations based on.

fixes https://github.com/redhat-performance/quads/issues/37

Change-Id: If74744bbdb9abcf0634b820b1f97f46423d710c2