Don't send notification when hosts are added later.
This fixes an issue with notifications where sometimes a
set of machines is later added to an existing assignment.

QUADS detects when the machine count changes and bases its
expiration notifications on this, but in cases where zero
machines are removed (but some are added) the user should not
receive notifications.  Additionally, no notification stub
should be created that marks that we diligently notified the
standard 7,5,3,1 days away from expiration.

Change-Id: Idb638e507187ec7f89451c2fa0d17f188587cdc2
fixes: https://github.com/redhat-performance/quads/issues/38