More precise checking on running process
The process check for quads-boot-order.sh tries to ensure that
not another process is running for the child process (ansible-playbook).
However, the test is to ensure that the PID for the running process
is valid.  This test in and of itself is not 100% reliable, as the PIDs
can be recycled on the system.  As such if some unrelated process is
running with the old PID for a previous run stored in the PID file,
quads will erroneously think there is another instance of the ansible
playbook process running.

The change to the pid lookup will limit the cmdline match search to
ensure the current running process is also "quads-boot-order" (or at
least includes it in the cmdline).

Change-Id: Iec349f98a1c95517c2aaa40027a4773e256b636f