Implement shellcheck for CI, minor change.
This implements "shellcheck" in test-quads.sh

I am also changing the interpreter to bash from
sh in bin/quads-boot-order.sh and removing reference
to the bin_dir variable which isn't used.

We will omit a lot of false positives initially but at the least it
should not pass CI if there are errors.

fixes: https://github.com/redhat-performance/quads/issues/41

Change-Id: Iab2abb00340afdc1db7d028188a6d0d4caaf191e