Further limit Ansible processes.
This puts in place thresholds to limit the Ansible
processes to a reasonable number.  These can get out
of hand when provisioning a few hundred hosts at once
and need to be run at a reasonable threshold.

This is done by making the subshell invoking the
ansible-playbook command searchable against the parent
tool name then limiting it to 60 concurrently running
instances.

fixes: https://github.com/redhat-performance/quads/issues/44

Change-Id: Ib9db6ce26119fd457f6c8e41eda9194ceaf2c00a