Make ansible_max_proc: a variable.
We recently implemented ansible-playbook process limitations
in order to manage system load better in response to a large
amount of provisioning actions occurring simultaneously:

https://github.com/redhat-performance/quads/commit/951f62afdeaadffa747f053a234a476f9864c6bd

This patch utilizes a variable for the maximum ansible process
count so it's more configurable.

ansible_max_proc: 60

We're setting 60 to be a sensible default.

I'm also updating the contributing part of the README.md
so that first-time usage of Gerrithub is better documented.

Change-Id: I17983aced4e5b6bdb5f48bf673459dc0b76243eb