calculate the next month correctly
There is a bug in the date command when on the months with 31 days
if you happen to be on day 31, the date calculations are done
incorrectly.  This fix is a but more cumbersome but will correctly
calculate the days months and years for the number of months out
we want to look.

Fixes: https://github.com/redhat-performance/quads/issues/53

Change-Id: I2a5509347861fe2d1c3d10e9e8f0480f5365f9ff