Release environment only after checks pass
The current notifications are not sent for the initial release
unless specific tests are passed with the environment.  The
initial pass at addressing this checks the QinQ and untagged
traffic for the given environment.  The changes to the notification
is coupled with a cron based check to ensure the env has passed
the initial network test.  Once it can be "released", the
releasability is marked using a combination of env-owner-ticket
marker (unique for every allocation).

Change-Id: I67fa8d2bf176df50b1f8947825525cdeffa24430
Fixes: https://github.com/redhat-performance/quads/issues/5