fix change introduced in --schedule-query
quads without any args produces the schedule
in yaml output.  the --schedule-query introduced
changed the default behavior.  This is a small
patch to change the default behavior and treats
the new arg as a toggle for producing the
schedule.   --month and --year are optional and
default to the current time's month and year.

Change-Id: I7ebd47eb1a84a4894a3a482ca6654e834f713914