Prevent --move-hosts with --date unless --dry-run
Looking at the code it appears it would be possible to
run --move-hosts with --date and we would actually trigger
a move of systems between clouds for a time other than
current.  This is dangerous and will lead to much anger
if it were to be done accidentally.

This patch prevents the use of --move-commands using --date
unless --dry-run is also done (causing a no-op; only a print
of what would happen).

Fixes: https://github.com/redhat-performance/quads/issues/63

Change-Id: I50a0c8fc831498b558851e172355afcbbae4122d