Fix time comparison when adding data
When adding clouds, the file on disk is
updated and hence becomes newer than the data because
the immediate history_init also causes a second write
which fails to find the file to be older than the load
data.  To fix this, we read the data back in immediately
after writing it.

Fixes: https://github.com/redhat-performance/quads/issues/72

Change-Id: Iceee5d0488024f66688277ad08b68e75ecf81759