Adding an extra name field to the instackenv.json
This is required for https://github.com/redhat-performance/quads/issues/13

Infrared automatically tags nodes with flavors based on this naame field
via https://github.com/redhat-openstack/infrared/blob/master/plugins/tripleo-overcloud/tag.yml

During automated deployments, we will have an external python script
that will designate a name to a node such as "r620compute-0",
"r620controller-1", "r730ceph-5" and write it out to the CSV so that it is
included in the instackenv.json. We will temporarily write "node" to the
instackenv.json in bin/make-instackenv-json.sh which will be replaced
with the name of the designated flavor when the python script runs later.
We are writing the dummy value "node" so that the list element
exists for the csv-to-instack.py script to run without errors.

Change-Id: I46480491b3677f63b417a5043a6ac57728c3ada8