Fix lock to only apply to move command
After running some tests, the number of read operations
that occur in the non-daemon mode of operation makes the
previous locking strategy too prohibitive.  When running
the "quads.py" command (standalone tool) in order to operate
on the schedule.yaml file, the race condition risk is most
impactful when running --move-hosts.  Therefore we move the
locking into the move_hosts class method.

Note that move_hosts is not used when running the quads-daemon
and using quads-cli.  Instead the CLI will use the API and
get pending_moves, and then the CLI will in turn run the
move command.  This avoids the race condition entirely as
there is only a single process that handles the schedule.yaml
file.  Furthermore the results are returned a lot quicker given
that we no longer open and close the schedule file for every read
and write operation.

Fixes: https://github.com/redhat-performance/quads/issues/109

Change-Id: Iec7195bff884851e5b1d3b4df406df0557c43e59