Ensure PXE is set on SuperMicro systems
These playbooks for the 1028r, 6018r, and 6048r are called
when hosts are marked for build and going from the "director"
build state into the "foreman" build state.  The playbook
calls out to ipmitool to ensure pxe is re-enabled on the target
host. This is typically necessary for applications that might
molest the boot order of the host (in this case director seems
to disable PXE on hosts when the host is in a deployed state).

When we reclaim the host, we want to enable PXE.

Fixes: https://github.com/redhat-performance/quads/issues/112

Change-Id: I0fe231ee06ed62c3dffa80a9d922ec25feae0357