Update QUADS tools for daemon/cli mode.
Now that we're using daemon/cli mode we should check
to make sure that the quads-daemon service is running.

This will add some sanity checking in each of our QUADS
tools to ensure that quads-daemon is indeed running before
we continue.

Additionally I'm adding a small change to our systemd unit
file to link back to our Github documentation.

fixes: https://github.com/redhat-performance/quads/issues/113

Change-Id: I4715e16a2321c8a5c29f7c7cdf8b085b586aebe5