Power Management
Currently a WIP. Need to get the ssh login/query peice sorted out. I can
currenly query all the outlets, using SSH. Need to have these mapped out
to specific hosts.

Outlets - New data structure to be used across vendor poewr strips.

Also, in the Host data structure, we will need to add the type of Power
Management, and the location -- this will have to be sorted out.

We might want to query the amount of power each strip is taking, we can
query this, however, to attempt to keep it generic, we might want to
create a "Strip" data struct where we store it. This could be easily
contained in a dict,however, I think it would be easier for other
implementations to create a object for bookkeeping.

Change-Id: I5a54bd4d43643d09d3f593fc2f58d77086fa2534