WIP: Power Management
This will utilize the pypi package I created called "Strips".

https://pypi.python.org/pypi/Strip

quads-cli and quads-daemon will utilize this package to manage the
Powerstrips.

Pulled in the PIP package "strip" directly into the quads/lib/
directory to avoid the need to have pip install done on the quads
host.  This library will need to be updated manually in the quads
repo, when the upstream PIP package is updated.

This is added as an optional feature in the QUADS host
setup documentation.

Fix some additional issues that were preventing proper startup.

Fixes: https://github.com/redhat-performance/quads/issues/100
Co-author: Kambiz Aghaiepour <kambiz@redhat.com>
Co-author: Will Foster <wfoster@redhat.com>

To define a strip:

   $quads --add-strip some-apc-01.example.com --brand Apc --username admin --password password

To (re)define a host with two power supplies:

   $quads  --define-host host04.example.com --default-cloud cloud01 --host-type vendor --force --strip-name some-apc-01.example.com,hostname.apc.com --strip-port A1,A3

Note that the PDU strips are comma separated, and the ports are also
comma separated (and the number in both lists should match otherwise
you will get an error.

Redefining a host that already exists preserves the existing details
including schedules.

To list strips:

   $quads --ls-strips

To get details on a specific strip:

   $quads --ls-strips --strip-only some-apc-01.example.com

To get a list of the outlets for a specific host:

   $quads --host host04.example.com --ls-outlets

Still need to be done via the CLI:

   --power-off
   --power-on
   --power-cycle

Change-Id: I5a54bd4d43643d09d3f593fc2f58d77086fa2534