Power Management
This will utilize the pypi package I created called "Strips".

https://pypi.python.org/pypi/Strip

quads-cli and quads-daemon will utilize this package to manage the
Powerstrips.

Pulled in the PIP package "strip" directly into the quads/lib/
directory to avoid the need to have pip install done on the quads
host.  This library will need to be updated manually in the quads
repo, when the upstream PIP package is updated.

Fixes: https://github.com/redhat-performance/quads/issues/100
Co-author: Kambiz Aghaiepour <kambiz@redhat.com>

Change-Id: I5a54bd4d43643d09d3f593fc2f58d77086fa2534