Add the ability to have host specific playbooks
In order to deal with hosts that may not confirm to the
same standard hardware configuration as the others of its
type, first check for:

   $install_dir/ansible/racadm-setup-boot-${target}-${playbook_type}.yml

where $playbook_type is either "foreman" or "director".  For example,
for a host named: b01-h01-r620.example.com we might not want to run:

   $install_dir/ansible/racadm-setup-boot-r620-foreman.yml

due to unavoidable hardware differences.  This change allows us to
create:

   $install_dir/ansible/racadm-setup-boot-b01-h01-r620.example.com-foreman.yml

which if found will be used instead of the more generic "$host_type"
playbooks.

Fixes: https://github.com/redhat-performance/quads/issues/120

Change-Id: Id47bdc056499625e5cfc777c4d81b306860b6ecb