Ensure default values when calling update_cloud
New paramaters were introduced to update_cloud in order
to do post configuration of specialized applications.
These are optional and need to default to None otherwise
the cloud definitions that don't pass them will fail.

Additionally --init is no longer required for quads-cli
and is performed the first time quads-daemon is invoked.

Fixes: https://github.com/redhat-performance/quads/issues/121

Change-Id: Ie55b725789e5e6df116bc9d7e29c9c8e629240d2