Implement systems level validation
This adds checks to ensure that all the hosts in a given
environment have completed their build state.  If there are
any hosts that have not finished building, then the network
validation tests are skipped and the error is reported.

Once the host builds are completed, then the network validation
tests are run.  Both have to pass for the environment to be
released to the user.

related-to: https://github.com/redhat-performance/quads/issues/25
fixes: https://github.com/redhat-performance/quads/issues/48

Change-Id: Ie6df8df17ffbe84e42ba63f895dcae5c3074c547