Add graceful shutdown attempt before poweraction.
This adds an attempt to gracefully shutdown the host OS
during foreman boot order playbook runs.  We do this once
but need to do it before attempting a hard reset.

Additionally we will increase the retries for larger
server classes (r730xd and r930) to allow for more time
in between lengthy reboots.

fixes: https://github.com/redhat-performance/quads/issues/131

Change-Id: I7c981e654449adc8664e7bbd615632f5689b6728