indexing validation bug
This issue was due to a lazy regex. The new regex will match IP
addresses and ignore them. Possibly we should use the QUADs config to
determine what will be in the hostname?

If there were multiple occurances of the same hostname caused issues.

This patch fixes both

Fixes bug: #135

Change-Id: I479d39760ff653f99302607b2024afaaacde3029