Add a color indicator when env is released
For the assignments page, this patch adds a green or red
spanned color to the text for the cloud name and the instackenv
link.  If the assignment is released, the instack link is provided
and the text is "download".  But if the assignment is not yet released,
the link points to a "underconstruction" URL on the quads host
(which will have to be manually configured by the installation) and
the text is changed to "validating".

Fixes: https://github.com/redhat-performance/quads/issues/138
Change-Id: I113555f89e95e726c18eb1be91c4cb37c664e656