Update patch set 37
Patch Set 37:

> (3 comments)
 > 
 > Joe,
 > 
 > We'll want to also include python-pip in rpm/quads.spec too since
 > you'll want people to pip install your strips.
 > 
 > 32   +Requires: python-pip >= 8.1

On second thought, let's not even use pip for this (but it's good it is available out there for others).

Why don't we just ship the Strip libs inside lib/ instead?  For maintenance we can just periodically backport it pip strips is your 'upstream'.

Only other comments:

1) can we change your license bits to GPLv3 to match rest of QUADS?

Patch-set: 37