Sync ansible playbooks
This small patch syncs the latest r630 playbooks.  The new
assumption is to enable PXE on all (both) interfaces that need
them, and then to just set the boot order to one of two possible
settings:

For foreman build:
  foreman nic,hard disk,director nic

For director ready hosts:
  director nic,hard disk,foreman nic

Also small fix for the not yet released environments to have a
"validating" text for the RED status lines in the assignments page.

Change-Id: I398617cf33bce3a7822c2438ed08484dafe53256