Simplify move and switchconf verify scripts
When the VLAN allocations per cloud environment
are configured to be sequential and have a formula
that can be applied, we no longer need to have
the associative arrays that would need to be updated
with each expansion of upper limits to supported
cloud numbers.

e.g. we use the following formula:

Each cloud has 4 vlans.  The clouds are defined as: "cloudXX"
XX is a number (e.g. your environment might support 01 through 24).

  offset = 1090 + (XX * 10)
  first_vlan = $offset
  last_vlan = $offset + 3

e.g. cloud01 would be :

  offset = 1090 + (01 * 10) = 1100
  first_vlan = 1100
  last_vlan = 1103

There is an offsets associative array that tells us
how to calculate the individual interface vlans. With:

  offsets=( ["em1"]="0" ["em2"]="1" ["em3"]="2" ["em4"]="3")

we calculate :

  vlan = offset (e.g. 1100) + (lookup offset for em1 through em4)

Another example:

  cloud04:

  offset = 1090 + (04 * 10) = 1130
  first_vlan = 1130
  last_vlan = 1133

  vlan for "em2" == offset (1130) + lookup em2 above (1) = 1131

Another example:

  cloud07.  Calculate VLAN for em4:

  offset = 1090 + (07 * 10) = 1160
  vlan for em4 = 1160 + 3 = 1163

https://github.com/redhat-performance/quads/issues/146

Change-Id: I4953d408adb4a8b37c86990eb8eb2674950439c4