Simplify move and switchconf verify scripts
When the VLAN allocations per cloud environment
are configured to be sequential and have a formula
that can be applied, we no longer need to have
the associative arrays that would need to be updated
with each expansion of upper limits to supported
cloud numbers.

e.g. we use the following formula:

Each cloud has 4 vlans.  The clouds are defined as: "cloudXX"
XX is a number (e.g. your environment might support 01 through 24).

  offset = 1090 + (XX * 10)
  first_vlan = $offset
  last_vlan = $offset + 3

e.g. cloud01 would be 1100 through 1103, cloud02 would be 1110 through
1113, etc.

Change-Id: I4953d408adb4a8b37c86990eb8eb2674950439c4