Add textures to visualization map
As the number of clouds increases, we need a better
way to differentiate between the environments when
visualizing them.  Previously we used simple colors
depicted using RGB values.  This patch allows for
a texture image to be used.

This also updates the RPM spec file and docs to
include httpd (or denotes setting up a web server
of the users choice if using git install method).

Lastly this adds an updated image example of the
validation status and ansible cmdb link from the
assignments page.

Co-Authored-By: Will Foster <wfoster@redhat.com>

Fixes: https://github.com/redhat-performance/quads/issues/151

Change-Id: Ib8cd0392435d09d2739a960831da335b886aad96