Add a way to preserve hosts when moving to new env
Fixes:  https://github.com/redhat-performance/quads/issues/165

wipe = 1 (or true) is still the default.  If we don't specify
it when we define a cloud, it is set to wipe resources it consumes.

However if we want to preserve hosts, we would define a cloud using

   quads-cli --define-cloud foo --description "foo" --wipe 0

We can see the current wipe state of the clouds using:

   quads --ls-wipe

Moves occur in the cli after getting the pending moves from the
quads-daemon.  Need to also lookup the wipe value of the destination
cloud to ensure proper wiping (or not wiping) as hosts are moved.

Change-Id: I2748c5d02dffe93312b7f9e85d8277a8d861156f