Enforce Foreman default medium/ptable with OS.
This enforces the standard Foreman partition table and OS medium
when we call --move-hosts command.  The reason for this is so that
if someone decides to reprovision their system with another Foreman
OS besides our current default (RHEL7) this will revert the
system settings to default again.

e.g.

Brunhilda uses Foreman to reprovision one of their servers to
RHEL8.  Later, when that server gets re-assigned to someone else
it may fail to reprovision properly.

fixes: https://github.com/redhat-performance/quads/issues/186

Change-Id: I513bd4a08629fe90986b48aa03cbb7e5065b4862