Allow moving a host even when config changes
The old way that move-and-rebuild-host.sh worked relied on
the juniper configs to be set a certain way and conform to the
templated format that looked something like:

  set interfaces $interfacename apply-groups QinQ_vl$vlan

The script would query the configuration and determine the old
vlan. However, it is not strictly necessary to know the old vlan.
Furthermore, if the switch configuration is changed to support
public IPs for the floating networks, then the method would fail
and the move of the hosts would also fail creating a cascading
failure that would require manual intervention.

This patch makes the attempt to determine the old vlan using the
same method as before (by querying the switch configuration), however
if it cannot determine it (e.g. when port is set for public floating
IP network), then it will assume that the old vlan passed on the command
line is correct, and calculates the vlan number accordingly.

Fixes:  https://github.com/redhat-performance/quads/issues/170

Change-Id: I30ccf780239c51ef421a4f1734c467395d194547