Adds an API that reports the future availability of all nodes in QUADS
Solves issue #194
Adds a new api call -show-reservations, which inquires the QUADS server
for all the nodes and its schedules and creates a report that shows
availability of each node grouped by its type and its availability in future

the api calculates  availability/busy status of group nodes based on the time it is invoked.
The status reported today by this api call  will be different then
the status tomorrow or few hours later.

It also provides a processed json file for further post processing.
Tenants can run this command and once they decide to schedule the nodes of particular type, they can then refer to the json file to decide on exact node names.