Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Hey Sahil, would you mind rebasing this on the latest set of changes?  We had to push a few patchsets that addressed immediate improvements for 1.0.
> 
> After rebasing and resolving conflicts when I try to start the quads server it gives me error: 
> ImportError: No module named lib.Foreman
> 
> To resolve it I tried installing python-foreman. That does not solve the issue. What module should I install to get this dependency.

That's from the Foreman host health checks patch, I have another patchset to opt-out of that (and will be the default) here when it gets merged:

https://review.gerrithub.io/c/redhat-performance/quads/+/432367

This should also fix testing/quads-sandbox.sh if you were using that to spin up a test environment.  @grafuls could comment on the best way to integrate the Foreman library otherwise if you wanted to include it.

Patch-set: 1