Add support for optional public VLANS
This covers augmenting quads-cli --define-cloud with an optional
argparse value like --extnet ${vlan_id} that creates a stub file
under ${quads_data}/vlans/ that holds up release of an assignment
until a public VLAN can be deployed on the 4th interface of the
hosts. Once this is done stub files need to be manually removed.

Validation is not sent if stub files are still present.

Co-authored-by: Kambiz Aghaiepour <kambiz@redhat.com>
Fixes: https://github.com/redhat-performance/quads/issues/192

Change-Id: I1fbff69e6917ac42fb5837ce0efd63c31ca99b19