Cleanup path issues and minor bug fix
Move all the PDU related items back into bin/ in order
to properly utilize yay.sh and load-config.sh without
replicating and creating unnecessary or confusing symlinks.

Also cleanup the bootstate/ file for hosts for which there is
no ansible playbooks in order to avoid creating and deleting
boot/ files when the expected bootstate/ is not thought to be
in sync with what it needs to be.

Change-Id: I3affd7d9412b394feb0552f3d0c3aaebf29b9eb9