Allow foreman where host entries are unified
In the original build of our environment, we incorrectly have
two distinct entries in foreman for hosts.  e.g.

    rackN-hostX-hostyype.domain
and
    mgmt-rackN-hostX-hostyype.domain

but the proper way to deal with hosts is to have a single entry
in foreman, and define the BMC interfaces as a secondary interface
for the host.  This allows us to properly take advantage of the
power management features in foreman assuming IPMI credentials are
also configured correctly.

In order for this to work, and the wiki content to generate properly
we rely instead of hammer, on the quads --ls-hosts output to generate
the initial host lists.  We also want to ensure the BMC interface is
not deleted when hosts are moved using our move-and-rebuild-hosts.sh

Change-Id: I801827589134711d21241b9376f1228e1aad4dc0