Split out CONTRIBUTING, Add Code of Conduct.
For our Github community guidelines (suggestions to make our project
more approachable) we should split out the contributing portions of
QUADS to a separate markdown.

https://github.com/redhat-performance/quads/community

There is also an opportunity to provide a simple code of conduct
for the project - something that mirrors the K.I.S.S. design and
simplicity we strive for but still serves to effectively illustrate our
ideal goals and conduct for the community.

In this way, the simple tenets derived from George Carlin's character
"Rufus" from 'Bill and Ted's Excellent Adventure' seem to fit the bill.

1)  Be excellent to each other
2)  Don't be bogus

One tenet here is to assume good intent in others, which makes the
simple duality of these two driving tenets possible without needing any
further clarification.

Change-Id: I32d3d6fbb825d224726aac6dc6620e1af274b240