Minor Doc Update:  CONTRIBUTING and build example.
Since we have split out the contributing documentation of our project we
should provide a very prominent link at the top of the page.

I'm also linking to our move-and-rebuild-host script as an example for
someone's move command.

Lastly I'm touching up the preparatory documentation and being more
specific about the needs and purpose of the QUADS move command.

Change-Id: I59111d02bd049c03e85ab3471224e5b08680024b