Add Foreman host health checks for add schedule.
Re-adding the ability for QUADS to check against Foreman for hosts
marked with the "broken_state: true" host parameter.  This allows you to
not schedule hosts which might be marked in a broken state.

foreman_check_host_health: true

By default we will leave this off, both for our CI (which doesn't bring
up a Foreman with hosts tied to QUADS) and development purposes but it
can be set by the QUADS user if they want it.

fixes: https://github.com/redhat-performance/quads/issues/211

Change-Id: Ie1801c4df6e6ff77ee48e44df137d5a2e59f79bc