Webkit ignores word-break CSS, extend wiki table.
Browsers that use Webkit like Google Chrome, Safari, do not properly
honor the word-break CSS attribute that we use to ensure long hostnames
are not wrapped. e.g.

.site-content {
       word-wrap: normal;
       word-break: keep-all;
}

A pragmatic fix to this is ensure our Markdown table column titles are
at least as long as our longest, unbroken string of text.

e.g. replacing ServerHostname with ServerHostnamePublic seems to do the
trick.

fixes: https://github.com/redhat-performance/quads/issues/216

Change-Id: I9a20902dd8801b1b91847d0b019f2f807eef2553