Fix Entropy Performance issue with libsodium, etc.
There seems to be a bug with entropy pool gathering with libsodium in
which calls to /dev/random fail.  This manifests itself in QEMU because
it has no hardware entropy (and as a result containers run in VMs).

As /dev/random and /dev/urandom are blocking devices this manifests
itself in slowdowns for random, sporadic calls from QUADS.  libsodium
which is used by paramiko cannot use /dev/urandom (non-blocking entropy)
so this moves us to use a software-based entropy service which is
optional.

https://github.com/redhat-performance/quads/issues/221

More about the nacl bindings / libsodium bug:
PyNACL / Sodium:

https://github.com/pyca/pynacl/issues/327

Symptoms / cause:

======
By default libsodium only uses the /dev/random path to determine if
the CSPRNG is initialized before it starts using /dev/urandom.
It also prefers to use getrandom, which gives the same behavior for
free (blocking only until the CSPRNG is ready). On normal systems this
happens extremely rapidly, but there exists a weird long tail of systems
that can have bad entropy for long periods of time
======

The pragmatic solution here is to instead use the haveged service to
provide software entropy.

* Make haveged a dependency on QUADS via RPM installation
* For containers include and start it in the Dockerfile

Included are other fixes/changes below:

* Add python3-ipdb to rpm spec requirements
* Specify version requirements for ipmitool and git
* Remove contactbank wp plugin, we're not using it and it may not even
work right.
* Since HTTPD is required for serving visuals/instackenv.json make it
also be a dependency of quads-server.service
* Demote docker compose from recommended status
* Correct RPM spec warnings about incorrect date format

Fixes: https://github.com/redhat-performance/quads/issues/221

Change-Id: I5c20defe12871e6399cf6b1ada659caf1a5e1b94