Fix for mod schedule on empty cloud.
Before if we were modifying a schedule date without passing a cloud
name, quads would strip out the cloud definition on the mongo document.
Fix this by passing already existing cloud if no cloud param is used
on the cli.
This patch also fixes redefine cloud as we were setting default values
for cloud document flags. Removed this and included on the field
definition instead.

Change-Id: I0484e0636c8dc672c8ecaa8652b248215ff3797c