Added cloud01 to wiki assignments summary.
When getting the summary for assignments we now rely on getting the
hosts via current_schedule method. Since we do not create any schedules
when moving to cloud01, the host count for this comes out as zero.
Added logic for getting host count for cloud01 by querying mongo host
document with cloud01 reference fields.

Fixes: https://github.com/redhat-performance/quads/issues/229

Change-Id: Ifc25e20a50c6f27fa38c810f6dae7a82e9d02cf6