Fix SMC badfish call, add tolerance to --move-hosts
This addresses a lack of exclusion of SuperMicro systems when running
the badfish library to lookup what would be a non-existent
idrac_interfaces.yml entry for "foreman" for one-time boot.  This is not
needed on SuperMicro systems.

This also removes the try/except here that would normally try to run the
badfish call against non-Dells (which will always fail) and instead just
doesn't try.

We're also changing the list of SUPERMICROS to SUPERMICRO in config.py
for no real reason other than cosmetic to match the method of
is_supermicro.

In bin/quads-cli we are removing the exit(1) when move_and_rebuild_hosts.py
as continue through the lists of hosts (but we can use return False
for portions of move_and_rebuild_hosts.py so these don't move).

Ultimately various stages of validation will catch issues but we don't
want move_and_rebuild_hosts.py to try over and over again (and fail) on
only one host of it's list.

fixes: https://github.com/redhat-performance/quads/issues/238
Change-Id: Ifd394f645419482c2a4247cf525888bc72e54dc9